@@ -53,20 +53,20 @@
     "publish:nrsrx-utils": "npm run build:nrsrx-utils && npm publish ./dist/libs/imng-nrsrx-client-utils",
     "dep-graph": "nx dep-graph",
     "help": "nx help",
-    "nx": "nx"
+    "nx": "nx",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/pwa": "^0.803.22",
-    "@angular/router": "^8.2.14",
-    "@angular/service-worker": "^8.2.14",
+    "@angular/animations": "^9.0.4",
+    "@angular/common": "^9.0.4",
+    "@angular/compiler": "^9.0.4",
+    "@angular/core": "^9.0.4",
+    "@angular/forms": "^9.0.4",
+    "@angular/platform-browser": "^9.0.4",
+    "@angular/platform-browser-dynamic": "^9.0.4",
+    "@angular/router": "^9.0.4",
+    "@angular/service-worker": "^9.0.4",
     "@fortawesome/angular-fontawesome": "^0.4.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.26",
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
@@ -75,7 +75,7 @@
     "@ngrx/entity": "8.6.0",
     "@ngrx/router-store": "8.6.0",
     "@ngrx/store": "^8.6.0",
-    "@nrwl/angular": "8.11.0",
+    "@nrwl/angular": "9.0.4",
     "@progress/kendo-angular-buttons": "^5.1.1",
     "@progress/kendo-angular-common": "^1.2.1",
     "@progress/kendo-angular-dateinputs": "^4.2.0",
@@ -98,38 +98,37 @@
     "ngx-bootstrap": "^5.2.0",
     "oidc-client": "^1.10.1",
     "rxjs": "~6.5.4",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.22",
-    "@angular-devkit/build-ng-packagr": "~0.803.22",
-    "@angular/cli": "^8.3.22",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.900.4",
+    "@angular-devkit/build-ng-packagr": "~0.900.4",
+    "@angular/cli": "^9.0.4",
+    "@angular/compiler-cli": "^9.0.4",
+    "@angular/language-service": "^9.0.4",
     "@ngrx/schematics": "8.6.0",
     "@ngrx/store-devtools": "8.6.0",
-    "@nrwl/cypress": "8.11.0",
-    "@nrwl/jest": "8.11.0",
-    "@nrwl/workspace": "8.11.0",
+    "@nrwl/cypress": "9.0.4",
+    "@nrwl/jest": "9.0.4",
+    "@nrwl/workspace": "9.0.4",
     "@types/jest": "24.0.9",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "copyfiles": "^2.1.1",
-    "cypress": "3.6.1",
+    "cypress": "3.8.3",
     "dotenv": "6.2.0",
     "jest": "24.1.0",
     "jest-junit": "^9.0.0",
-    "jest-preset-angular": "^7.1.1",
-    "ng-packagr": "^5.7.1",
+    "jest-preset-angular": "8.0.0",
+    "ng-packagr": "^9.0.0",
     "ngrx-store-freeze": "0.2.4",
     "odata-ts-generator": "^0.1.28",
     "prettier": "1.19.1",
     "rimraf": "^2.7.1",
     "ts-jest": "24.0.0",
     "ts-node": "~7.0.0",
-    "tsickle": "^0.35.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.5.3"
+    "typescript": "~3.7.5"
   }
 }
